KRACOEUS-6956 : repack propdev.impl.person.creditsplit
