+++ coeus-code/src/main/java/org/kuali/coeus/propdev/impl/person/ProposalPerson.java
@@ -32,8 +32,8 @@ import org.kuali.kra.award.home.ContactRole;
 import org.kuali.kra.bo.AbstractProjectPerson;
 import org.kuali.kra.bo.CitizenshipType;
 import org.kuali.kra.budget.personnel.PersonRolodex;
-import org.kuali.kra.proposaldevelopment.bo.CreditSplitNameInfo;
-import org.kuali.kra.proposaldevelopment.bo.CreditSplitable;
+import org.kuali.coeus.propdev.impl.person.creditsplit.CreditSplitNameInfo;
+import org.kuali.coeus.propdev.impl.person.creditsplit.CreditSplitable;
 import org.kuali.kra.proposaldevelopment.bo.DevelopmentProposal;
 import org.kuali.kra.proposaldevelopment.bo.ProposalPersonCreditSplit;
 import org.kuali.coeus.propdev.impl.person.question.ProposalPersonQuestionnaireHelper;
