+++ coeus-code/src/main/java/org/kuali/kra/award/contacts/AwardPersonUnitCreditSplitRule.java
@@ -15,7 +15,7 @@
  */
 package org.kuali.kra.award.contacts;
 
-import org.kuali.kra.proposaldevelopment.bo.InvestigatorCreditType;
+import org.kuali.coeus.common.framework.type.InvestigatorCreditType;
 
 /**
  * AwardCreditSplitRule
