+++ coeus-code/src/main/java/org/kuali/kra/institutionalproposal/contacts/InstitutionalProposalPersonUnitCreditSplitRuleImpl.java
@@ -16,7 +16,7 @@
 package org.kuali.kra.institutionalproposal.contacts;
 
 import org.kuali.kra.institutionalproposal.rules.InstitutionalProposalCreditSplitAuditError;
-import org.kuali.kra.proposaldevelopment.bo.InvestigatorCreditType;
+import org.kuali.coeus.common.framework.type.InvestigatorCreditType;
 import org.kuali.coeus.sys.framework.rule.KcTransactionalDocumentRuleBase;
 import org.kuali.coeus.sys.api.model.ScaleTwoDecimal;
