+++ coeus-code/src/main/resources/org/kuali/kra/institutionalproposal/repository-institutionalproposal.xml
@@ -442,7 +442,7 @@
 		<field-descriptor name="objectId" column="OBJ_ID" jdbc-type="VARCHAR" />
 		<field-descriptor name="versionNumber" column="VER_NBR" jdbc-type="BIGINT" locking="false" />
 
-    	<reference-descriptor name="investigatorCreditType" class-ref="org.kuali.kra.proposaldevelopment.bo.InvestigatorCreditType">
+    	<reference-descriptor name="investigatorCreditType" class-ref="org.kuali.coeus.common.framework.type.InvestigatorCreditType">
 			<foreignkey field-ref="invCreditTypeCode" />
 		</reference-descriptor>
 		<reference-descriptor name="institutionalProposalPerson" class-ref="org.kuali.kra.institutionalproposal.contacts.InstitutionalProposalPerson" auto-retrieve="true" auto-update="none" auto-delete="none"> 
@@ -491,7 +491,7 @@
         <reference-descriptor name="institutionalProposal" class-ref="org.kuali.kra.institutionalproposal.home.InstitutionalProposal" auto-retrieve="false" auto-update="none" auto-delete="none"> 
             <foreignkey field-ref="proposalId"/>
         </reference-descriptor>
-        <reference-descriptor name="investigatorCreditType" class-ref="org.kuali.kra.proposaldevelopment.bo.InvestigatorCreditType" auto-retrieve="false" auto-update="none" auto-delete="none"> 
+        <reference-descriptor name="investigatorCreditType" class-ref="org.kuali.coeus.common.framework.type.InvestigatorCreditType" auto-retrieve="false" auto-update="none" auto-delete="none">
             <foreignkey field-ref="invCreditTypeCode"/>
         </reference-descriptor>
     </class-descriptor> 
@@ -511,7 +511,7 @@
 		<field-descriptor name="objectId" column="OBJ_ID" jdbc-type="VARCHAR" />
 		<field-descriptor name="versionNumber" column="VER_NBR" jdbc-type="BIGINT" locking="false" />
 
-    	<reference-descriptor name="investigatorCreditType" class-ref="org.kuali.kra.proposaldevelopment.bo.InvestigatorCreditType">
+    	<reference-descriptor name="investigatorCreditType" class-ref="org.kuali.coeus.common.framework.type.InvestigatorCreditType">
 			<foreignkey field-ref="invCreditTypeCode" />
 		</reference-descriptor>
 		<reference-descriptor name="institutionalProposalPersonUnit" class-ref="org.kuali.kra.institutionalproposal.contacts.InstitutionalProposalPersonUnit" auto-retrieve="true" auto-update="none" auto-delete="none"> 
@@ -677,7 +677,7 @@
 		<reference-descriptor name="institutionalProposal" class-ref="org.kuali.kra.institutionalproposal.home.institutionalProposal" auto-retrieve="false" auto-update="none" auto-delete="none"> 
 			<foreignkey field-ref="proposalId"/>
 		</reference-descriptor>
-		<reference-descriptor name="investigatorCreditType" class-ref="org.kuali.kra.proposaldevelopment.bo.InvestigatorCreditType" auto-retrieve="false" auto-update="none" auto-delete="none"> 
+		<reference-descriptor name="investigatorCreditType" class-ref="org.kuali.coeus.common.framework.type.InvestigatorCreditType" auto-retrieve="false" auto-update="none" auto-delete="none">
 			<foreignkey field-ref="invCreditTypeCode"/>
 		</reference-descriptor>
 	</class-descriptor>
